Add XML comments to the LoggingHook
